+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog and release notes
 
+## Unreleased
+### Fixes
+- fix throwing error when `of => objectType` wasn't provided in abstract resolver class
+
 ## v0.12.2
 ### Features
 - add support for using type classes in browser (configure webpack to use decorators shim)